Add Firebase Info.plist / Clean up Project
