+++ SupportingFiles/GoogleService/GoogleService-Info.plist
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>CLIENT_ID</key>
+	<string>549644893019-bbonhvjvkjjelm00e2sbplhhhq69036g.apps.googleusercontent.com</string>
+	<key>REVERSED_CLIENT_ID</key>
+	<string>com.googleusercontent.apps.549644893019-bbonhvjvkjjelm00e2sbplhhhq69036g</string>
+	<key>API_KEY</key>
+	<string>AIzaSyCOUOH4mEIMHNgTdhWVWnAL9vaxkYHa7KI</string>
+	<key>GCM_SENDER_ID</key>
+	<string>549644893019</string>
+	<key>PLIST_VERSION</key>
+	<string>1</string>
+	<key>BUNDLE_ID</key>
+	<string>com.dayolstudio.Dayol.beta</string>
+	<key>PROJECT_ID</key>
+	<string>dayol-beta</string>
+	<key>STORAGE_BUCKET</key>
+	<string>dayol-beta.appspot.com</string>
+	<key>IS_ADS_ENABLED</key>
+	<false></false>
+	<key>IS_ANALYTICS_ENABLED</key>
+	<false></false>
+	<key>IS_APPINVITE_ENABLED</key>
+	<true></true>
+	<key>IS_GCM_ENABLED</key>
+	<true></true>
+	<key>IS_SIGNIN_ENABLED</key>
+	<true></true>
+	<key>GOOGLE_APP_ID</key>
+	<string>1:549644893019:ios:ab753cf0ea83afde4db87e</string>
+</dict>
+</plist>
